Cleaned up warnings

+++ commands/rmdir.c
@@ -23,10 +23,16 @@ int wmain(int argc, wchar_t *argv[]){
 		i++;
 	}
 
+	if (argv[1] == 0){
+		return EXIT_FAILURE;
+	}
+
 	if (RemoveDirectory(argv[1]) == 0) {
 		wprintf(L"\nCouldn't remove %s directory.\n", argv[1]);
 	}
 	else {
 		wprintf(L"\n%s directory removed successfully.\n", argv[1]);
 	}
+
+	return EXIT_SUCCESS;
 }
